chore: Split replicache.test.ts
It was getting too large. This breaks out all the tests that contains
'subscribe'/'subscription' in their title.
